+++ pkg/vm/errors.go
@@ -3,6 +3,7 @@ package vm
 import "fmt"
 
 var (
-	NotOperable  = fmt.Errorf("not operable")
-	NotIndexable = fmt.Errorf("not indexable")
+	NotOperable   = fmt.Errorf("not operable")
+	NotIndexable  = fmt.Errorf("not indexable")
+	NotComparable = fmt.Errorf("not comparable")
 )